Extend the Rails frameworks correctly
As the code was before, this gem was forcing the Active Record and
Active Model frameworks to be loaded too early in the Rails boot process.

Instead of Active Record and Active Model being loaded when the the
application stats to initialize, they would be loaded when this gem is
required, making the Rails configuration options for these frameworks
to not be applied from the application.